check-certificates credhub logic to scripts/lib
re-use later
